+++ crates/edr_eth/Cargo.toml
@@ -7,50 +7,38 @@ edition = "2021"
 anyhow = "1.0.75"
 alloy-rlp = { version = "0.3", default-features = false, features = ["derive"] }
 c-kzg = { version = "1.0.0", default-features = false }
-futures = { version = "0.3.28", default-features = false }
 hash-db = { version = "0.15.2", default-features = false }
 hash256-std-hasher = { version = "0.15.2", default-features = false }
 hashbrown = { version = "0.14.3", default-features = false, features = ["ahash", "allocator-api2", "inline-more"] }
 hex = { version = "0.4.3", default-features = false, features = ["alloc"] }
-hyper = { version = "0.14.27", default-features = false }
 itertools = { version = "0.10.5", default-features = false, features = ["use_alloc"] }
 k256 = { version = "0.13.1", default-features = false, features = ["arithmetic", "ecdsa", "pkcs8", ] }
-lazy_static = { version = "1.4.0", default-features = false }
 log = { version = "0.4.17", default-features = false }
-reqwest = { version = "0.11", features = ["blocking", "json"] }
-reqwest-middleware = { version = "0.2.4", default-features = false }
-reqwest-retry = { version = "0.3.0", default-features = false }
-reqwest-tracing = { version = "0.4.7", default-features = false, optional = true }
 revm-primitives = { git = "https://github.com/NomicFoundation/revm", rev = "aceb093", version = "3.1", default-features = false, features = ["c-kzg", "hashbrown"] }
 serde = { version = "1.0.147", default-features = false, features = ["derive"], optional = true }
-serde_json = { version = "1.0.89", optional = true }
 sha2 = { version = "0.10.8", default-features = false }
 sha3 = { version = "0.10.8", default-features = false }
 thiserror = { version = "1.0.37", default-features = false }
-tokio = { version = "1.21.2", default-features = false, features = ["fs", "macros", "sync"] }
 tracing = { version = "0.1.37", features = ["attributes", "std"], optional = true }
 triehash = { version = "0.8.4", default-features = false }
 uuid = { version = "1.4.1", default-features = false, features = ["v4"] }
-url = "2.4.1"
-regex = "1.10.0"
 
 [dev-dependencies]
 anyhow = "1.0.75"
 assert-json-diff = "2.0.2"
 edr_defaults = { version = "0.3.5", path = "../edr_defaults" }
-lazy_static = "1.4.0"
-mockito = { version = "1.0.2", default-features = false }
-paste = { version = "1.0.14", default-features = false }
 edr_test_utils = { version = "0.3.5", path = "../edr_test_utils" }
+lazy_static = "1.4.0"
+paste = { version = "1.0.14", default-features = false }
+serde_json = { version = "1.0.89" }
 serial_test = "2.0.0"
-tempfile = { version = "3.7.1", default-features = false }
 tokio = { version = "1.23.0", features = ["macros"] }
 walkdir = { version = "2.3.3", default-features = false }
 
 [features]
 default = ["std"]
 rand = ["revm-primitives/rand"]
-serde = ["dep:serde", "c-kzg/serde", "revm-primitives/serde", "serde_json"]
-std = ["futures/std", "hash256-std-hasher/std", "hash-db/std", "hex/std", "itertools/use_std", "k256/std", "k256/precomputed-tables", "revm-primitives/std", "serde?/std", "sha2/std", "sha3/std", "triehash/std", "uuid/std"]
-tracing = ["dep:tracing", "reqwest-tracing"]
+serde = ["dep:serde", "c-kzg/serde", "revm-primitives/serde"]
+std = ["hash256-std-hasher/std", "hash-db/std", "hex/std", "itertools/use_std", "k256/std", "k256/precomputed-tables", "revm-primitives/std", "serde?/std", "sha2/std", "sha3/std", "triehash/std", "uuid/std"]
+tracing = ["dep:tracing"]
 test-remote = ["serde"]
