+++ crates/edr_rpc_client/Cargo.toml
@@ -1,0 +1,30 @@
+[package]
+name = "edr_rpc_client"
+version = "0.3.5"
+edition = "2021"
+
+[dependencies]
+anyhow = { version = "1.0.75", default-features = false, features = ["std"] }
+edr_eth = { version = "0.3.5", path = "../edr_eth", features = ["serde"] }
+futures = { version = "0.3.28", default-features = false, features = ["std"] }
+hyper = { version = "0.14.27", default-features = false }
+lazy_static = { version = "1.4.0", default-features = false }
+log = { version = "0.4.17", default-features = false }
+regex = "1.10.0"
+reqwest = { version = "0.11", default-features = false, features = ["blocking", "json"] }
+reqwest-middleware = { version = "0.2.4", default-features = false }
+reqwest-retry = { version = "0.3.0", default-features = false }
+reqwest-tracing = { version = "0.4.7", default-features = false, optional = true }
+serde = { version = "1.0.147", default-features = false, features = ["derive", "std"] }
+serde_json = { version = "1.0.89" }
+thiserror = { version = "1.0.37", default-features = false }
+tokio = { version = "1.21.2", default-features = false, features = ["fs", "macros", "sync"] }
+tracing = { version = "0.1.37", default-features = false, features = ["attributes", "std"], optional = true }
+url = { version = "2.4.1", default-features = false }
+
+[dev-dependencies]
+mockito = { version = "1.0.2", default-features = false }
+tempfile = { version = "3.7.1", default-features = false }
+
+[features]
+tracing = ["dep:tracing", "dep:reqwest-tracing"]
