+++ crates/edr_rpc_eth/src/lib.rs
@@ -1,19 +1,17 @@
 mod block_spec;
 mod cacheable_method_invocation;
+/// Input type for `eth_call` and `eth_estimateGas`
+mod call_request;
 mod chain_id;
-/// an Ethereum JSON-RPC client
-pub mod client;
 /// ethereum objects as specifically used in the JSON-RPC interface
 pub mod eth;
 /// data types for use with filter-based RPC methods
 pub mod filter;
-/// data types specific to JSON-RPC but not specific to Ethereum.
-pub mod jsonrpc;
 mod r#override;
 mod request_methods;
 
 pub use self::{
     block_spec::{BlockSpec, BlockTag, Eip1898BlockSpec, PreEip1898BlockSpec},
-    client::{RpcClient, RpcClientError},
+    call_request::CallRequest,
     r#override::*,
 };
